bookstore/pyproject.toml
[tool.poetry]
name = "bookstore"
package-mode = false
version = "0.1.0"
description = "Bookstore API"
authors = ["lmescoito <lmescoito@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
factory-boy = "^3.3.0"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
